Use gsize instead of gint. From Miroslaw Dobrzanski-Neumann
2002-04-08  Sebastian Wilhelmi  <wilhelmi@ira.uka.de>

	* tests/gio-test.c, tests/markup-test.c: Use gsize instead of
	gint. From Miroslaw Dobrzanski-Neumann
	<mne@mosaic-ag.com>. (#77982)

+++ tests/gio-test.c
@@ -64,7 +64,7 @@ read_all (int         fd,
 	  guint      *bytes_read)
 {
   guint left = nbytes;
-  guint nb;
+  gsize nb;
   GIOError error = G_IO_ERROR_NONE;
   char *bufp = buffer;
